requirements.txt
opencv-python==4.5.5.64
Pillow==9.3.0
numpy==1.21.5
scipy==1.7.3
scikit-image==0.19.2
scikit-learn==1.0.2
tensorflow==2.8.0
keras==2.8.0
ImageHash==4.2.1
pandas==1.3.5
openpyxl==3.0.9
futures==3.0.5
logging==0.5.1.2
requests==2.26.0
tqdm==4.62.3